+++ R/a_freq_j.R
@@ -4,6 +4,8 @@
 #' and (optional) relative risk columns
 #'
 #' @inheritParams proposal_argument_convention
+#' @param .var (`character(1)`)\cr Name of a categorical analysis variable in `df`.
+#' The variable must be a `factor`.
 #' @param val (`character` or NULL)\cr
 #' When NULL, all levels of the incoming variable (variable used in the `analyze` call)
 #' will be considered.\cr
@@ -94,6 +96,7 @@ s_freq_j <- function(
     stop("Argument .var cannot be NA or NULL.")
   }
 
+  checkmate::assert_string(.var)
   checkmate::assert_string(id)
   checkmate::assert_subset(id, colnames(df), empty.ok = FALSE)
 
@@ -102,6 +105,9 @@ s_freq_j <- function(
   if (countsource %in% c("altdf", "altdf_subset")) {
     df <- alt_df
   }
+
+  checkmate::assert_names(names(df), must.include = .var)
+  checkmate::assert_class(df[[.var]], classes = "factor")
 
   .alt_df <- alt_df